+++ apps/client/next.config.js
@@ -3,11 +3,6 @@ module.exports = {
   reactStrictMode: false,
   transpilePackages: ['@repo/ui'],
   images: {
-    domains: [
-      'example.com',
-      'another-domain.com',
-      'www.naver.com',
-      'adaptors-bucket.s3.ap-northeast-2.amazonaws.com',
-    ], // 허용할 도메인 추가
+    domains: ['adaptors-bucket.s3.ap-northeast-2.amazonaws.com'],
   },
 };
